The gyro code should now be in its final state for the third robot.
Conflicts:
	bot3/input/gyro_reader.cc
	frc971/atom_code/atom_code.gyp
	frc971/atom_code/build.sh
	frc971/crio/build.sh
	frc971/input/input.gyp

+++ aos/common/control_loop/Timing.cpp
@@ -11,8 +11,8 @@ namespace time {
 void PhasedLoopXMS(int ms, int offset) {
   // TODO(brians): Tests!
   const Time frequency = Time::InMS(ms);
-  SleepUntil((Time::Now() / frequency.ToNSec()) *
-             frequency.ToNSec() +
+  SleepUntil((Time::Now() / static_cast<int32_t>(frequency.ToNSec())) *
+             static_cast<int32_t>(frequency.ToNSec()) +
              frequency + Time::InUS(offset));
 }
